feat: add LottoController class
